Only admin log into parent planet (fixes #572) (#578)

+++ src/app/home/home.component.ts
@@ -117,10 +117,13 @@ export class HomeComponent implements OnInit, AfterViewInit, OnDestroy {
 
   logoutClick() {
     this.userService.endSessionLog().pipe(switchMap(() => {
-      return  forkJoin([
-        this.couchService.delete('_session', { withCredentials: true }),
-        this.couchService.delete('_session', { withCredentials: true, domain: this.userService.getConfig().parent_domain }),
-      ]);
+      const obsArr = [ this.couchService.delete('_session', { withCredentials: true }) ];
+      if (this.userService.getConfig().name === this.userService.get().name) {
+        obsArr.push(
+          this.couchService.delete('_session', { withCredentials: true, domain: this.userService.getConfig().parent_domain }),
+        );
+      }
+      return forkJoin(obsArr);
     })).subscribe((response: any) => {
         this.userService.unset();
         this.router.navigate([ '/login' ], {});
